Migrated to match API to V5

+++ app/src/main/java/com/tsarsprocket/reportmid/model/ItemModel.kt
@@ -3,7 +3,7 @@ package com.tsarsprocket.reportmid.model
 import com.tsarsprocket.reportmid.RIOTIconProvider
 
 class ItemModel(
-    val riotId: Long,
+    val riotId: Int,
     val name: String,
     imageName: String,
     iconProvider: RIOTIconProvider,
